--- a/.clang-format
+++ b/.clang-format
@@ -1,0 +1,1 @@
+{ BasedOnStyle: Google, IndentWidth: 4 }