--- a/.gitignore
+++ b/.gitignore
@@ -56,6 +56,7 @@ TAGS
 # other
 .DS_Store
 
+/runtime/include
 _build/
 _build.*/
 *.agdai
@@ -69,6 +70,8 @@ UPDATES-FOR-CHANGELOG.org
 .juvix-build
 
 # C Code generation
+*.out.c
+*.s
 *.wasm
 docs/md/
 _docs