--- a/app/Commands/Dev/Runtime/Compile.hs
+++ b/app/Commands/Dev/Runtime/Compile.hs
@@ -119,6 +119,7 @@ commonArgs o outputFile =
     <> [ "-W",
          "-Wall",
          "-Wno-unused-parameter",
+         "-Wno-unused-label",
          "-Werror",
          "-std=c11",
          "-I",