--- a/app/Commands/Dev/Runtime/Options.hs
+++ b/app/Commands/Dev/Runtime/Options.hs
@@ -1,0 +1,24 @@
+module Commands.Dev.Runtime.Options where
+
+import Commands.Dev.Runtime.Compile.Options
+import CommonOptions
+
+newtype RuntimeCommand
+  = Compile RuntimeCompileOptions
+  deriving stock (Data)
+
+parseRuntimeCommand :: Parser RuntimeCommand
+parseRuntimeCommand =
+  hsubparser $
+    mconcat
+      [ commandCompile
+      ]
+  where
+    commandCompile :: Mod CommandFields RuntimeCommand
+    commandCompile = command "compile" compileInfo
+
+    compileInfo :: ParserInfo RuntimeCommand
+    compileInfo =
+      info
+        (Compile <$> parseRuntimeCompileOptions)
+        (progDesc "Compile a C file with Juvix runtime included")