--- a/docs/org/notes/runtime-benchmark-results.org
+++ b/docs/org/notes/runtime-benchmark-results.org
@@ -1,7 +1,7 @@
 Benchmarks of the new Juvix runtime
 ====================================
 
-Benchmarked version: commit 148ececb4d4259eacbb980f5992073a3ac611d82 from 31.11.2022
+Benchmarked version: commit 148ececb4d4259eacbb980f5992073a3ac611d82 from 31.10.2022
 
 Summary
 -------
@@ -51,7 +51,7 @@ In addition to the `fibonacci` and `combinations` benchmarks, the
 advantage of direct transpilation for very simple programs is also
 visible on the `fold` benchmark where a simple loop over a list
 dominates the running time. However, this is partly because the
-compilation of closures in current Juvix is incorrect which allows it
+compilation of closures in current Juvix is incorrect allowing it
 to be more efficient.
 
 Benchmark programs