--- a/runtime/Makefile
+++ b/runtime/Makefile
@@ -14,7 +14,7 @@ native32:
 	$(MAKE) -f Makefile.generic CONFIG=NATIVE32+RELEASE
 
 native64:
-	$(MAKE) -f Makefile.generic CONFIG=NATIVE64+RELEASE
+	make -f Makefile.generic CONFIG=NATIVE64+RELEASE
 
 x86_32:
 	$(MAKE) -f Makefile.generic CONFIG=X86_32+RELEASE
@@ -29,7 +29,7 @@ native32-debug:
 	$(MAKE) -f Makefile.generic CONFIG=NATIVE32+DEBUG
 
 native64-debug:
-	$(MAKE) -f Makefile.generic CONFIG=NATIVE64+DEBUG
+	make -f Makefile.generic CONFIG=NATIVE64+DEBUG
 
 x86_32-debug:
 	$(MAKE) -f Makefile.generic CONFIG=X86_32+DEBUG