--- a/runtime/Makefile
+++ b/runtime/Makefile
@@ -52,4 +52,4 @@ clean:
 	@-rm -rf _build*
 	@-rm -rf include
 
-.PHONY: release debug includes format wasm32 wasm32-wasi native32 native64 x86_32 wasm32-debug wasm32-wasi-debug native32-debug native64-debug x86_32-debug asm clean
+.PHONY: release debug includes wasm32 wasm32-wasi native32 native64 x86_32 wasm32-debug wasm32-wasi-debug native32-debug native64-debug x86_32-debug asm format clean