--- a/runtime/Makefile
+++ b/runtime/Makefile
@@ -1,8 +1,8 @@
 all: release debug
 
-release: wasm32-wasi native32
+release: wasm32-wasi native64
 
-debug: wasm32-wasi-debug native32-debug
+debug: wasm32-wasi-debug native64-debug
 
 wasm32:
 	$(MAKE) -f Makefile.generic CONFIG=WASM32+RELEASE
@@ -12,6 +12,9 @@ wasm32-wasi:
 
 native32:
 	$(MAKE) -f Makefile.generic CONFIG=NATIVE32+RELEASE
+
+native64:
+	$(MAKE) -f Makefile.generic CONFIG=NATIVE64+RELEASE
 
 x86_32:
 	$(MAKE) -f Makefile.generic CONFIG=X86_32+RELEASE
@@ -25,10 +28,13 @@ wasm32-wasi-debug:
 native32-debug:
 	$(MAKE) -f Makefile.generic CONFIG=NATIVE32+DEBUG
 
+native64-debug:
+	$(MAKE) -f Makefile.generic CONFIG=NATIVE64+DEBUG
+
 x86_32-debug:
 	$(MAKE) -f Makefile.generic CONFIG=X86_32+DEBUG
 
 clean:
 	@-rm -rf _build*
 
-.PHONY: release debug wasm32 wasm32-wasi native32 x86_32 wasm32-debug wasm32-wasi-debug native32-debug x86_32-debug clean
+.PHONY: release debug wasm32 wasm32-wasi native32 native64 x86_32 wasm32-debug wasm32-wasi-debug native32-debug native64-debug x86_32-debug clean