--- a/runtime/Makefile
+++ b/runtime/Makefile
@@ -34,7 +34,10 @@ native64-debug:
 x86_32-debug:
 	$(MAKE) -f Makefile.generic CONFIG=X86_32+DEBUG
 
+format:
+	clang-format -i `find src -name '*.c' -or -name '*.h'`
+
 clean:
 	@-rm -rf _build*
 
-.PHONY: release debug wasm32 wasm32-wasi native32 native64 x86_32 wasm32-debug wasm32-wasi-debug native32-debug native64-debug x86_32-debug clean
+.PHONY: release debug format wasm32 wasm32-wasi native32 native64 x86_32 wasm32-debug wasm32-wasi-debug native32-debug native64-debug x86_32-debug clean