--- a/tests/runtime/positive/out/test001.out
+++ b/tests/runtime/positive/out/test001.out
@@ -1,0 +1,1 @@
+Hello world!