# Description

Basic version of the C runtime necessary to run compiled Juvix programs.

Closes #1554

# Checklist:

- [x] Page allocator
- [x] Allocator for unstructured objects
- [x] Global stack
- [x] Structured object interface:
  - [x] Header word specification with macros to query & manipulate the bit-encoded fields
  - [x] Constructor allocator
- [x] Integers implemented by unboxed 31-bit ints, at first without overflow check
- [x] Constant strings
- [x] Function call interface:
  - [x] Macros for calling functions, saving & restoring live variables, allocating & extending closures
  - [x] Trampoline (dispatch loop) to implement JuvixAsm CallClosures
- [x] Structural equality on runtime values
- [x] Static storage of information about functions and constructors (e.g. for pretty-printing)
- [x] Pretty-printing of runtime values
- [x] Stacktraces
- [x] Interpreter for IO actions
- [x] API with macros directly corresponding to JuvixReg instructions
- [x] Makefile to create static libraries with different configurations (debug, release, wasm32-anoma, wasm32-cosm, native32-libc, native64-libc) 
- [x] Tests
- [x] Benchmarks