@lukaszcz 2022-10-31
Benchmarks of the new Juvix runtime
====================================

Benchmarked version: commit 148ececb4d4259eacbb980f5992073a3ac611d82 from 31.11.2022

Summary
-------

We benchmark several programs manually compiled into the primitives of
the new Juvix runtime. The code corresponds closely to the code that
will be generated by the new compilation process, with basic low-level
optimisations (unboxing, untagging, etc.) but without any high-level
optimisations on JuvixCore (inlining, specialisation, constant
folding, fusion, etc.). This corresponds to the compilation process
planned for the 0.4 milestone.

We compare the running time and memory usage with analogous programs
written in Haskell, OCaml, JuvixCore (using the evaluator), current
Juvix (with the "direct" transpilation to C) and C.

The results suggest that for most first-order programs the new
compilation process will produce code with running time comparable to
the code produced by the native OCaml compiler. For higher-order
programs heavy on closure manipulation, the results are acceptable but
noticeably worse, especially with third-order functions
(i.e. functions which take functions taking functions). This could,
however, be alleviated by implementing the specialisation optimisation
(see the "specialised" column in the `ackermann` and `mapfun`
benchmarks). Besides, functional programs of order higher than two are
rare.

The comparisons with OCaml and Haskell were not entirely fair because
the new Juvix runtime does not perform garbage collection. The
overhead of garbage collection is particularly visible on the
`mergesort` benchmark which creates many intermediate data structures
that are quickly discarded. With proper memory management, the running
time results on first-order programs for the new Juvix runtime are
expected to become slightly worse than for the native OCaml compiler.

For simple programs operating on integers which don't require any heap
memory allocation (`fibonacci` and `combinations` benchmarks), the
direct transpilation to C in the current Juvix seems to perform best
(behind only C). The reason is that for very simple programs `clang`
can better optimise the output of such a direct transpiler. The main
problem with the transpilation to C approach is that it cannot scale
to reliably work for more complex programs, as evidenced by the
segfaults, longer running time and higher memory use on other
benchmarks.

In addition to the `fibonacci` and `combinations` benchmarks, the
advantage of direct transpilation for very simple programs is also
visible on the `fold` benchmark where a simple loop over a list
dominates the running time. However, this is partly because the
compilation of closures in current Juvix is incorrect which allows it
to be more efficient.

Benchmark programs
------------------

* fibonacci: compute the Nth Fibonacci number modulo 2^28 (N = 100’000’000)

    The Nth Fibonacci number is computed in O(N). Needs only constant
    stack space and no heap memory. This benchmark tests the
    efficiency of tail recursion and arithmetic operations.

* combinations: count combinations of numbers 1 to N having sum N (N = 100)

    This benchmark tests the efficiency of general recursion. No heap
    memory needs to be allocated. Uses stack space proportional to
    N. The running time is exponential in N.

* prime: compute the Nth prime (N = 16384)

    The Nth prime number is computed via the Eratosthenes sieve. A
    list of N primes is created. No intermediate lists are discarded
    (garbage collection not needed). This benchmark tests the
    efficiency of tail recursion, arithmetic operations, list cell
    allocation and access.

* mergesort: merge sort a list of N integers (N = 2’000’000)

    At each level of merge sort intermediate lists are created and
    discarded. The running time for this benchmark largely depends on
    the efficiency of memory management. Here one may observe the
    overhead of garbage collection or the memory blow-up if no garbage
    collection is used.

* maybe: optionally sum N integers from a binary tree K times (N = 2^20, K = 100)

    If a fixed number k is encountered in the tree then the result is
    `Nothing`, otherwise it is `Just sum`. The computation is repeated
    for values of k from 0 to K. This tests the efficiency of handling
    optional values and data structure access.

* fold: fold a list of N integers K times (N = 100’000, K = 1000)

    The sum of N natural numbers is computed via fold_left
    (tail-recursive). The computation is repeated K times. The list is
    created only once, so that allocation time does not dominate. This
    benchmark tests the efficiency of closure call and list cell
    access.

* cps: compute the Nth Fibonacci number modulo 2^28 with CPS (N = 100’000’000)

    The function computing the Nth Fibonacci number is written in
    continuation-passing style, tail-recursively calling a
    continuation supplied as an argument. This benchmark tests the
    efficiency of closure call and allocation.

* mapfold: map and fold a list of N integers K times (N = 10000, K = 10000)

    This benchmark tests the efficiency of standard higher-order
    functions on lists, closure call and memory management. The
    program allocates O(K) intermediate lists of length N which are
    quickly discarded.

* ackermann: compute Ack(3, N) with the higher-order Ackermann function definition (N = 11)

    The higher-order Ackermann function definition iterates an
    iteration of function compositions. Hence, it uses a third-order
    invocation of an iteration function. This benchmark tests the
    efficiency of creating and calling second-order closures, and of
    partial application.

* mapfun: successively map K functions to a list of N integers (K = 100, N = 10000)

    The benchmark stores K second-order closures in a list, maps them
    successively to a list of K closures, and then successively maps
    the K closures from the result to a list of N integers. This
    benchmark tests the efficiency of manipulating closures and
    storing them in data structures.

The benchmark programs can be found in `tests/benchmark` in the Juvix
source directory.

Methodology
-----------

For each program the total running time (elapsed real time) and memory
use (maximum resident set size) were measured on an M1 iMac with no
significant background activity. Averages of several runs were taken.
The variance was negligible, unless indicated otherwise by providing a
range.

Results
-------

* fibonacci: compute the Nth Fibonacci number modulo 2^28 (N = 100’000’000)

|                        |New Juvix runtime (native)|New Juvix runtime (wasm32, wasmer)|Current Juvix (native)|Current Juvix (wasm32, wasmer)|JuvixCore evaluator|Haskell (native, ghc -O2)|Haskell (native, ghc -XStrict -O2)|OCaml (native, ocamlopt -O2)|OCaml (bytecode)|C (native, clang -O3)|C (wasm32, clang -Os, wasmer)|
|------------------------|--------------------------|----------------------------------|----------------------|------------------------------|-------------------|-------------------------|----------------------------------|----------------------------|----------------|---------------------|-----------------------------|
|Time (seconds, real)    |0.26                      |0.35                              |0.35                  |0.23                          |13.15              |10.03                    |0.39                              |0.35                        |0.94            |0.16                 |0.22                         |
|Memory use (MB, max RSS)|1.5                       |3.8                               |1.3                   |8.8                           |21.3               |8067.7                   |9.7                               |1.7                         |1.8             |1.3                  |4.0                          |

* combinations: count all combinations of numbers 1 to N having sum N (N = 1000)

|                        |New Juvix runtime (native)|New Juvix runtime (wasm32, wasmer)|Current Juvix (native)|Current Juvix (wasm32, wasmer)|JuvixCore evaluator|Haskell (native, ghc -O2)|Haskell (native, ghc -XStrict -O2)|OCaml (native, ocamlopt -O2)|OCaml (bytecode)|C (native, clang -O3)|C (wasm32, clang -Os, wasmer)|
|------------------------|--------------------------|----------------------------------|----------------------|------------------------------|-------------------|-------------------------|----------------------------------|----------------------------|----------------|---------------------|-----------------------------|
|Time (seconds, real)    |6.67                      |11.25                             |3.22                  |5.1                           |441.71             |5.48                     |5.48                              |6.53                        |41.08           |2.69                 |4.80                         |
|Memory use (MB, max RSS)|1.5                       |3.9                               |1.3                   |8.9                           |22.3               |9.6                      |9.6                               |1.7                         |1.9             |1.3                  |4.0                          |


* prime: compute the Nth prime (N = 16384)

|                        |New Juvix runtime (native)|New Juvix runtime (wasm32, wasmer)|Current Juvix (native)|Current Juvix (wasm32, wasmer)|JuvixCore evaluator|Haskell (native, ghc -O2)|Haskell (native, ghc -XStrict -O2)|OCaml (native, ocamlopt -O2)|OCaml (bytecode)|C (native, clang -O3)|C (wasm32, clang -Os, wasmer)|
|------------------------|--------------------------|----------------------------------|----------------------|------------------------------|-------------------|-------------------------|----------------------------------|----------------------------|----------------|---------------------|-----------------------------|
|Time (seconds, real)    |1.52                      |1.91                              |segfault              |3.09                          |167.04             |3.85                     |3.85                              |1.68                        |14.82           |0.12                 |0.13                         |
|Memory use (MB, max RSS)|1.7                       |4.0                               |segfault              |9.3                           |24.4               |9.8                      |9.6                               |2.2                         |2.2             |1.4                  |4.0                          |


* mergesort: merge sort a list of N integers (N = 2’000’000)

|                        |New Juvix runtime (native)|New Juvix runtime (wasm32, wasmer)|Current Juvix (native)|Current Juvix (wasm32, wasmer)|JuvixCore evaluator|Haskell (native, ghc -O2)|Haskell (native, ghc -XStrict -O2)|OCaml (native, ocamlopt -O2)|OCaml (bytecode)|C (native, clang -O3)|C (wasm32, clang -Os, wasmer)|
|------------------------|--------------------------|----------------------------------|----------------------|------------------------------|-------------------|-------------------------|----------------------------------|----------------------------|----------------|---------------------|-----------------------------|
|Time (seconds, real)    |0.40                      |0.31                              |3.55                  |1.32                          |22.45              |2.86                     |2.90                              |1.95                        |3.52            |0.15                 |0.15                         |
|Memory use (MB, max RSS)|1973.7                    |720.4                             |5046.7                |2729.8                        |1728.9             |253.6                    |253.6                             |172.6                       |343.1           |24.4                 |26.8                         |


* maybe: optionally sum N non-zero integers from a binary tree K times (N = 2^20, K = 100)

|                        |New Juvix runtime (native)|New Juvix runtime (wasm32, wasmer)|Current Juvix (native)|Current Juvix (wasm32, wasmer)|JuvixCore evaluator|Haskell (native, ghc -O2)|Haskell (native, ghc -XStrict -O2)|OCaml (native, ocamlopt -O2)|OCaml (bytecode)|C (native, clang -O3)|C (wasm32, clang -Os, wasmer)|
|------------------------|--------------------------|----------------------------------|----------------------|------------------------------|-------------------|-------------------------|----------------------------------|----------------------------|----------------|---------------------|-----------------------------|
|Time (seconds, real)    |0.45                      |0.64                              |3.29                  |1.57                          |22.75              |5.58                     |0.59                              |0.30                        |3.57            |0.27                 |0.50                         |
|Memory use (MB, max RSS)|1.6                       |3.8                               |2646.1                |1320.9                        |22.4               |5560.7                   |9.7                               |3.9                         |4.0             |1.3                  |4.1                          |

* fold: fold a list of N integers K times (N = 100’000, K = 1000)

|                        |New Juvix runtime (native)|New Juvix runtime (wasm32, wasmer)|Current Juvix (native)|Current Juvix (wasm32, wasmer)|JuvixCore evaluator|Haskell (native, ghc -O2)|Haskell (native, ghc -XStrict -O2)|OCaml (native, ocamlopt -O2)|OCaml (bytecode)|C (native, clang -O3)|C (wasm32, clang -Os, wasmer)|
|------------------------|--------------------------|----------------------------------|----------------------|------------------------------|-------------------|-------------------------|----------------------------------|----------------------------|----------------|---------------------|-----------------------------|
|Time (seconds, real)    |0.45                      |0.54                              |0.35                  |0.23                          |15.27              |0.58                     |0.58                              |0.36                        |1.80            |NA                   |NA                           |
|Memory use (MB, max RSS)|3.1                       |4.6                               |4.4                   |10.6                          |43.4               |12.7                     |12.7                              |5.9                         |5.9             |NA                   |NA                           |


* cps: compute the Nth Fibonacci number modulo 2^28 with CPS (N = 100’000’000)

|                        |New Juvix runtime (native)|New Juvix runtime (wasm32, wasmer)|Current Juvix (native)|Current Juvix (wasm32, wasmer)|JuvixCore evaluator|Haskell (native, ghc -O2)|Haskell (native, ghc -XStrict -O2)|OCaml (native, ocamlopt -O2)|OCaml (bytecode)|C (native, clang -O3)|C (wasm32, clang -Os, wasmer)|
|------------------------|--------------------------|----------------------------------|----------------------|------------------------------|-------------------|-------------------------|----------------------------------|----------------------------|----------------|---------------------|-----------------------------|
|Time (seconds, real)    |0.43                      |0.52                              |1.56                  |stack overflow                |20.22              |10.04                    |0.39                              |0.35                        |1.60            |0.16                 |0.25                         |
|Memory use (MB, max RSS)|1.5                       |3.9                               |1539.3                |stack overflow                |21.3               |8067.7                   |9.7                               |1.7                         |1.8             |1.3                  |4.0                          |

* mapfold: map and fold a list of N integers K times (N = 10000, K = 10000)

|                        |New Juvix runtime (native)|New Juvix runtime (wasm32, wasmer)|Current Juvix (native)|Current Juvix (wasm32, wasmer)|JuvixCore evaluator|Haskell (native, ghc -O2)|Haskell (native, ghc -XStrict -O2)|OCaml (native, ocamlopt -O2)|OCaml (bytecode)|C (native, clang -O3)|C (wasm32, clang -Os, wasmer)|
|------------------------|--------------------------|----------------------------------|----------------------|------------------------------|-------------------|-------------------------|----------------------------------|----------------------------|----------------|---------------------|-----------------------------|
|Time (seconds, real)    |1.01                      |1.59                              |2.74                  |1.81                          |38.24              |1.29                     |2.42                              |1.43                        |4.22            |NA                   |NA                           |
|Memory use (MB, max RSS)|2154.5                    |893.0                             |3059.1                |1542.0                        |26.4               |10.6                     |10.7                              |7.5                         |10-20           |NA                   |NA                           |

* ackermann: compute Ack(3, N) with the higher-order Ackermann function definition (N = 11)

|                        |New Juvix runtime (native)|New Juvix runtime (wasm32, wasmer)|New Juvix runtime (specialised, native)|New Juvix runtime (specialised, wasm32, wasmer)|Current Juvix (native)|Current Juvix (wasm32, wasmer)|JuvixCore evaluator|Haskell (native, ghc -O2)|Haskell (native, ghc -XStrict -O2)|OCaml (native, ocamlopt -O2)|OCaml (bytecode)|C (native, clang -O3)|C (wasm32, clang -Os, wasmer)|
|------------------------|--------------------------|----------------------------------|---------------------------------------|-----------------------------------------------|----------------------|------------------------------|-------------------|-------------------------|----------------------------------|----------------------------|----------------|---------------------|-----------------------------|
|Time (seconds, real)    |0.92                      |1.21                              |0.30                                   |0.65                                           |segfault              |runtime error                 |11.71              |0.87                     |0.47                              |0.54                        |1.35            |0.00                 |0.14                         |
|Memory use (MB, max RSS)|2.6                       |4.1                               |2.3                                    |3.9                                            |segfault              |runtime error                 |23.3               |13.6                     |9.6                               |2.0                         |3.6             |1.3                  |4.0                          |

* mapfun: successively map K functions to a list of N integers (K = 100, N = 10000)

|                        |New Juvix runtime (native)|New Juvix runtime (wasm32, wasmer)|New Juvix runtime (specialised, native)|New Juvix runtime (specialised, wasm32, wasmer)|Current Juvix (native)|Current Juvix (wasm32, wasmer)|JuvixCore evaluator|Haskell (native, ghc -O2)|Haskell (native, ghc -XStrict -O2)|OCaml (native, ocamlopt -O2)|OCaml (bytecode)|C (native, clang -O3)|C (wasm32, clang -Os, wasmer)|
…